Added a timer and music for attacks, and attacks
The timer and the music have their timing synchronized to a BPM field
for player. The potential distortion caused by adjusting `pitch_scale`
directly has been mitigated with a pitch shift effect tied to the music
audio bus.

This bus will shift the pitch, but not the playback speed, whereas the
pitch shift for the audio playback will change both. So, we can use it
to correct for distortions from accelerated playback.