Added organ, organized beat code better.
The organ is a layered instrument, using the new
`AudioStreamSynchronized` stream type. It can be independently muted if
necessary.

The tempo needs to be consistent, but we can trigger different elements
of play based on different tracks. That way, each element of the game
(enemies, weapons, etc.) can associate with a different sound.